Commit 2: A New Hope
scraped old method, The new method will consist of a Empty Object with
both on and off versions of the self illuminated object the on version
will also have the actual Unity Light as a subobject. The Empty Object
(with the sub objects) has a script attached to change the power state
and set the avtive state of each object this should be controlled by a
room script so the power distrabution terminal will work when
implmented.
+FloodLight script
+FloodLight prefab
+FloodLight On Object
+FloodLight Off Object