8: Core Inventory Finish
Replaced some of the placeholder icons with the actual ones, saved the
inventory and options objects as a prefab for main game placement. (set
up requiered :package: ). Wrote Swap and Drop methods to actually work
and with the given buttons. Hooked up the e button to open the main
inventory, adjusted the selected item code so when hand held items are
this will be fairly easy to do. Added text to display the number of an
item in the top right conorner for ease this also became the button to
open the options. Added a giveRandomItem funtion to give a random item
with an ammount between min and max. Adjusted all functions to deal with
all major changes.
ItemSwitching.cs
+toggleInventory() switches the main inventory on and off
+swap() swaps the selected item to a selected position
+drop() drops one of the selected item
+updateText() updates the text displaying the ammount of the item
Item Text prefab: used as the defualt color and font settings for the
text that display the number of items
options.prefab: two buttons swap and drop to appear when an item is
clicked in the inventory.