2: Options Scale and Position
Changed the options to appear above the inventory as opposed to on the
item itself. Will need to adjust Inventory to be more clear for user
interaction later as a result tho. Updated the prefabs to reflect
changes, moved prefab items into a better folder. Deleted legacy prefab.
ItemSwitching.cs:
+Added static start postion of options
-Removed postion of options by selected object